updated betterw warnings, optimization with Release mode

+++ src/getline/getline.c
@@ -3,7 +3,7 @@
 #include<string.h>
 #include<stdlib.h>
 
-char *f_getline(char buffer[static 1], size_t bytes, FILE *fptr) {
+char *f_getline(char buffer[static 1], int bytes, FILE *fptr) {
 	buffer[0] = 0;
 	char *ret = fgets(buffer, bytes, fptr);
 	if(ret) {
