+++ src/getline/getline.h
@@ -1,5 +1,5 @@
 #pragma once
 #include<stdio.h>
 
-extern char *f_getline(char buffer[static 1], size_t bytes, FILE *fptr);
+extern char *f_getline(char buffer[static 1], int bytes, FILE *fptr);
